+++ examples/vite-demo/tsconfig.eslint.json
@@ -13,7 +13,10 @@
     "noUnusedParameters": true,
     "noImplicitReturns": true,
     "jsx": "preserve",
-    "jsxImportSource": "solid-js",
+    "jsxImportSource": "solid-js",,
+    "types": [
+      "solid-marked"
+    ]
   },
   "include": ["./src"]
 }
