+++ examples/vite-demo/tsconfig.json
@@ -14,6 +14,9 @@
     "noImplicitReturns": true,
     "jsx": "preserve",
     "jsxImportSource": "solid-js",
+    "types": [
+      "solid-marked"
+    ]
   },
   "include": ["./src"]
 }
