+++ packages/solid-marked/compiler/index.ts
@@ -1,5 +1,3 @@
-// eslint-disable-next-line @typescript-eslint/triple-slash-reference
-/// <reference path="../env.d.ts" />
 import {
   RawSourceMap,
   SourceNode,
