+++ packages/solid-marked/src/index.ts
@@ -1,3 +1,5 @@
+// eslint-disable-next-line @typescript-eslint/triple-slash-reference
+/// <reference path="global-types.d.ts" />
 import {
   createComponent,
   createContext,
